backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
pydantic[email]>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=3.0.0,<4.0.0
python-multipart>=0.0.6
pytest>=7.4.0
httpx>=0.24.0
fastapi-mail>=1.4.0
python-dotenv>=1.0.0
authlib>=1.3.0
